Introducing progress bars.
The proa.py main window now has two progress bars. These are used
during resets to notify the user that there is still something
going on. The first one basically reports the current line number
in the code. The second one depends on the part of the reset
that's currently being run. In findjobs.gimmeAll, it shows the
directory count, in findjobs.openFile it shows the data being
parsed, and for network transfers of findjobs.openFile results,
it shows the raw byte count.